requirements.txt
aiogram<3.0
environs~=8.0.0
asyncio==3.4.3
requests==2.26.0
beautifulsoup4==4.9.3
lxml==4.6.3
pandas==1.3.2
XlsxWriter==3.0.1
fake-headers==1.0.2
PyExifTool==0.5.2